remove graphql module
