+++ src/app.module.ts
@@ -1,8 +1,5 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { GraphQLModule } from '@nestjs/graphql';
-import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo';
-import { join } from 'path';
 import { JobModule } from './job/job.module';
 import ormConfig from './config/orm.config';
 import { TypeOrmModule } from '@nestjs/typeorm';
@@ -16,13 +13,6 @@ import { TypeOrmModule } from '@nestjs/typeorm';
       expandVariables: true,
     }),
     TypeOrmModule.forRoot(ormConfig()),
-    GraphQLModule.forRoot<ApolloDriverConfig>({
-      driver: ApolloDriver,
-      autoSchemaFile: join(process.cwd(), 'src/schema.gql'),
-      subscriptions: {
-        'graphql-ws': true,
-      },
-    }),
     JobModule,
   ],
 })
