--- src/job/pubsub.ts
@@ -1,2 +1,0 @@
-import { PubSub } from 'graphql-subscriptions';
-export const pubSub = new PubSub();
