--- src/schema.gql
@@ -1,37 +1,0 @@
-# ------------------------------------------------------
-# THIS FILE WAS AUTOMATICALLY GENERATED (DO NOT MODIFY)
-# ------------------------------------------------------
-
-type Job {
-  id: Int!
-  name: String!
-  status: JobStatus!
-  createdAt: DateTime!
-}
-
-"""The status of a job"""
-enum JobStatus {
-  PENDING
-  IN_PROGRESS
-  COMPLETED
-  FAILED
-}
-
-"""
-A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format.
-"""
-scalar DateTime
-
-type Query {
-  jobs: [Job!]!
-}
-
-type Mutation {
-  createJob(name: String!): Job!
-  updateJobStatus(id: Int!, status: String!): Job!
-}
-
-type Subscription {
-  jobAdded: Job!
-  jobUpdated: Job!
-}
